add link permission

+++ poolink_backend/apps/link/api/views.py
@@ -70,21 +70,26 @@ class LinkView(BaseAPIView):
     )
     @permission_classes([IsWriterOrReadonly])
     def post(self, request):
-        serializer = LinkSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            favicon = Favicon().get_favicon(serializer.validated_data['url'])
+        if request.user == Board.objects.get(id=request.data['board']).user:
+            print("유저 동일")
+            serializer = LinkSerializer(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                favicon = Favicon().get_favicon(serializer.validated_data['url'])
 
-            Link.objects.create(
-                board=serializer.validated_data['board'],
-                label=serializer.validated_data['label'],
-                url=serializer.validated_data['url'],
-                show=serializer.validated_data['show'],
-                favicon=favicon
-            )
-            return Response(
-                status=HTTP_200_OK,
-                data=MessageSerializer({"message": _("링크를 저장했습니다.")}).data,
-            )
+                Link.objects.create(
+                    board=serializer.validated_data['board'],
+                    label=serializer.validated_data['label'],
+                    url=serializer.validated_data['url'],
+                    show=serializer.validated_data['show'],
+                    favicon=favicon
+                )
+                return Response(
+                    status=HTTP_200_OK,
+                    data=MessageSerializer({"message": _("링크를 저장했습니다.")}).data,
+                )
+        else:
+            print("유저 다름")
+            return Response(data=MessageSerializer({"message": _("접근 권한이 없습니다.")}).data)
 
     @swagger_auto_schema(
         operation_id=_("Delete My Link"),
@@ -95,14 +100,18 @@ class LinkView(BaseAPIView):
     )
     @permission_classes([IsWriterOrReadonly])
     def delete(self, request):
-        serializer = LinkDestroySerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            Link.objects.filter(
-                board_id__in=Board.objects.filter(user=request.user).values('id'),
-                id__in=serializer.validated_data["links"],
-            ).delete()
+        for i in range(len(request.data['links'])):
+            if request.user != Link.objects.get(id=request.data['links'][i]).board.user:
+                print(i, "번째 링크의 소유자는 다른 유저이기 때문에 삭제에서 제외됩니다.")
+                continue
+            serializer = LinkDestroySerializer(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                Link.objects.filter(
+                    board_id__in=Board.objects.filter(user=request.user).values('id'),
+                    id__in=serializer.validated_data["links"],
+                ).delete()
 
-        return Response(status=HTTP_204_NO_CONTENT, data=MessageSerializer({"message": _("링크를 삭제했습니다.")}).data)
+            return Response(status=HTTP_204_NO_CONTENT, data=MessageSerializer({"message": _("링크를 삭제했습니다.")}).data)
 
 
 link_view = LinkView.as_view()
